chore(blade): convert static pages from wordpress theme to blade
This branch is for ongoing work using Laravel for the front end discussed in #17. Static pages are currently working on Homestead! Still much to do. Final goal is for all front-end views to be rendered by Laravel while the database and Admin tasks are still managed by Wordpress which is installed in a subdirectorty