feature(webapp): Direct installed webapps to SOTD by default
The homepage is really meant for new folks, but if someone has added the webapp
to their homescreen, they are most likely a regular/returning user and the
homepage is not the best spot for them to end up by default. Better to just go
straight to Scripture of the Day.

WEB-219 #done